+++ packages/gasket-typescript-tests/test/plugin-service-worker.spec.ts
@@ -12,9 +12,7 @@ describe('@gasket/plugin-service-worker', () => {
           maxAge: 1000 * 60
         },
         scope: '/docs',
-        minify: {
-          ie8: true
-        }
+        minify: true
       }
     };
   });